refactor dpm campaign workflow routes
