+++ src/app/router_registry.py
@@ -20,6 +20,9 @@ from app.routers.dpm_proof_packs import router as dpm_proof_packs_router
 from app.routers.dpm_wave_campaign_definitions import (
     router as dpm_wave_campaign_definitions_router,
 )
+from app.routers.dpm_wave_campaign_workflow import (
+    router as dpm_wave_campaign_workflow_router,
+)
 from app.routers.dpm_waves import router as dpm_waves_router
 from app.routers.foundation import router as foundation_router
 from app.routers.intake import router as intake_router
@@ -57,6 +60,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_outcome_reviews_router)
     app.include_router(dpm_command_center_pm_quality_router)
     app.include_router(dpm_wave_campaign_definitions_router)
+    app.include_router(dpm_wave_campaign_workflow_router)
     app.include_router(dpm_construction_router)
     app.include_router(dpm_proof_packs_router)
     app.include_router(dpm_waves_router)
